pyproject.toml
[project]
name = "network-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncssh>=2.21.1",
    "fastapi>=0.122.0",
    "genie>=25.10",
    "jinja2>=3.1.6",
    "pyats>=25.10",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scrapli-community>=2025.1.30",
    "scrapli[asyncssh,genie,ssh2,textfsm]>=2025.1.30",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
asyncssh = [
    "scrapli>=2025.1.30",
]
